Added label, increment, decrement, and length component
Added label component that will be used with break and session label.

Added increment component that will be used with break and session label.

Added decrement component that will be used with break and session label.

Added partial length component that will be with break and session.
Have to add value props for default value.