Created Break and Session values decrement code
Created the methods (decreaseBreak and decreaseSession) and implementation in React to decrease the values.

Created the validation so that neither the Break nor Session value can go below 0.

Removed the decrement component and just added divs to reduce complexity of the code.

Added a min-width to the break and session values for design.